# resource_management/resource_management.service/resource_management.service.client/src/main/resources/api-docs-for-ui.yaml
...
openapi: 3.0.1
info:
  title: Resource Management
  description: REST API of Resource Management
  contact:
    name: Fraunhofer IPA
    url: https://www.ipa.fraunhofer.de
    email: matthias.schneider@ipa.fraunhofer.de
  version: 1.4.0-SNAPSHOT
servers:
  - url: https://slm.local/resource-management
    description: Generated server url
security:
  - bearer_auth: []
  - spring_oauth: []
paths:
...
  "/resources/{resourceId}/remote-access":
    put:
      tags:
        - Resources
      summary: Set remote access of resource with id
      operationId: setRemoteAccessOfResource
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: resourceUsername
          in: query
          required: false
          schema:
            type: string
        - name: resourcePassword
          in: query
          required: false
          schema:
            type: string
        - name: resourceConnectionType
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/ConnectionType"
        - name: resourceConnectionPort
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/resources/{resourceId}/location":
    put:
      tags:
        - Resources
      summary: Set location of resource with id
      operationId: setLocationOfResource
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: locationId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/resources/locations/{locationId}":
    put:
      tags:
        - Location
      summary: Add new Location with id
      operationId: addLocation
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Location"
  "/resources":
    get:
      tags:
        - Resources
      summary: Get all resources
      operationId: getResources
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/BasicResource"
    post:
      tags:
        - Resources
      summary: Add existing resource
      operationId: addExistingResource
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateResourceRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
                format: uuid
  "/resources/{resourceId}/submodels":
    get:
      tags:
        - Resources AAS
      summary: Get resource submodels
      operationId: getResourceSubmodels
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
    post:
      tags:
        - Resources AAS
      summary: Add submodels to existing resource
      operationId: addSubmodels
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - aasx
              type: object
              properties:
                aasx:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/resources/locations":
    get:
      tags:
        - Location
      summary: Get locations
      operationId: getLocations
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Location"
    post:
      tags:
        - Location
      summary: Add new Location
      operationId: addLocation_1
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Location"
    delete:
      tags:
        - Location
      summary: Delete Location
      operationId: deleteLocation
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  "/importer/file":
    post:
      tags:
        - Importer
      summary: Import from file
      operationId: importFromFile
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
...
  "/resources/{resourceId}/submodels/{submodelIdBase64Encoded}":
    delete:
      tags:
        - Resources AAS
      summary: Delete resource submodel
      operationId: deleteSubmodel
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: submodelIdBase64Encoded
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
components:
  schemas:
    CreateResourceRequest:
      type: object
      properties:
        resourceHostname:
          type: string
        resourceIp:
          type: string
        digitalNameplateV3:
          "$ref": "#/components/schemas/DigitalNameplateV3"
    DigitalNameplateV3:
      required:
        - addressInformation
        - manufacturerName
        - manufacturerProductDesignation
        - uriOfTheProduct
      type: object
      properties:
        uriOfTheProduct:
          type: string
        manufacturerName:
          type: string
        manufacturerProductDesignation:
          type: string
        addressInformation:
          type: string
        manufacturerProductRoot:
          type: string
        manufacturerProductFamily:
          type: string
        manufacturerProductType:
          type: string
        orderCodeOfManufacturer:
          type: string
        productArticleNumberOfManufacturer:
          type: string
        serialNumber:
          type: string
        yearOfConstruction:
          type: string
        dateOfManufacture:
          type: string
          format: date-time
        hardwareVersion:
          type: string
        firmwareVersion:
          type: string
        softwareVersion:
          type: string
        countryOfOrigin:
          type: string
        uniqueFacilityIdentifier:
          type: string
    ConnectionType:
      type: string
      enum:
        - ssh
        - WinSsh
        - WinRM
        - http
        - tcp
    Location:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    Action:
      required:
        - actionClass
        - configParameters
        - connectionTypes
        - skipable
      type: object
      properties:
        actionClass:
          type: string
        actionType:
          "$ref": "#/components/schemas/ActionType"
        skipable:
          type: boolean
        configParameters:
          type: array
          items:
            "$ref": "#/components/schemas/ActionConfigParameter"
        connectionTypes:
          uniqueItems: true
          type: array
          items:
            "$ref": "#/components/schemas/ConnectionType"
      discriminator:
        propertyName: actionClass
    ActionConfigParameter:
      required:
        - description
        - name
        - prettyName
        - requiredType
        - secret
        - valueType
      type: object
      properties:
        name:
          type: string
        prettyName:
          type: string
        description:
          type: string
        valueType:
          "$ref": "#/components/schemas/ActionConfigParameterValueType"
        defaultValue:
          type: string
        requiredType:
          "$ref": "#/components/schemas/ActionConfigParameterRequiredType"
        secret:
          type: boolean
    ActionConfigParameterRequiredType:
      type: string
      enum:
        - ALWAYS
        - SKIP
    ActionConfigParameterValueType:
      type: string
      enum:
        - SERVICE_PORT
        - STRING
        - FILE
        - KUBE_CONF
    ActionType:
      type: string
      enum:
        - INSTALL
        - UNINSTALL
        - SCALE_UP
        - SCALE_DOWN
        - CREATE_VM
        - DELETE_VM
        - DEPLOY
        - UNDEPLOY
        - UPDATE
        - RUN
    AwxAction:
      required:
        - actionClass
        - awxBranch
        - awxRepo
        - configParameters
        - connectionTypes
        - playbook
        - skipable
      type: object
      allOf:
        - "$ref": "#/components/schemas/Action"
        - type: object
          properties:
            awxRepo:
              type: string
            awxBranch:
              type: string
            playbook:
              type: string
            parameter:
              type: array
              items:
                "$ref": "#/components/schemas/SurveyItem"
...
    SurveyItem:
      required:
        - choices
        - default
        - question_description
        - question_name
        - required
        - type
        - variable
      type: object
      properties:
        type:
          "$ref": "#/components/schemas/SurveyItemType"
        question_name:
          type: string
        question_description:
          type: string
        variable:
          type: string
        choices:
          type: array
          items:
            type: string
        min:
          type: integer
          format: int32
        max:
          type: integer
          format: int32
        required:
          type: boolean
        default:
          type: string
    SurveyItemType:
      type: string
      enum:
        - text
        - password
        - integer
        - float
        - multiplechoice
        - multiselect
...
    CapabilityHealthCheck:
      required:
        - interval
        - path
        - port
        - type
      type: object
      properties:
        type:
          "$ref": "#/components/schemas/CapabilityHealthCheckType"
        port:
          type: integer
          format: int32
        path:
          type: string
        interval:
          type: integer
          format: int32
    CapabilityHealthCheckType:
      type: string
      enum:
        - HTTP
    CapabilityType:
      type: string
      enum:
        - SETUP
        - DEPLOY
        - SCALE
        - VM
    ClusterMemberType:
      required:
        - minNumber
        - name
        - prettyName
        - scalable
      type: object
      properties:
        name:
          type: string
        prettyName:
          type: string
        minNumber:
          type: integer
          format: int32
        scalable:
          type: boolean
...
    DeploymentType:
      type: string
      enum:
        - DOCKER_CONTAINER
        - DOCKER_COMPOSE
        - KUBERNETES
        - CODESYS
    ExecutionEnvironment:
      required:
        - description
        - image
        - pull
      type: object
      properties:
        image:
          type: string
        pull:
          "$ref": "#/components/schemas/ExecutionEnvironmentPull"
        description:
          type: string
        registryCredential:
          "$ref": "#/components/schemas/RegistryCredentials"
    ExecutionEnvironmentPull:
      type: string
      enum:
        - ''
        - always
        - missing
        - never
    RegistryCredentials:
      required:
        - authenticationURL
        - password
        - username
        - verifySSL
      type: object
      properties:
        id:
          type: integer
          format: int32
        description:
          type: string
        authenticationURL:
          type: string
        username:
          type: string
        password:
          type: string
        verifySSL:
          type: boolean
...
    BaseConfigurationCapability:
      required:
        - actions
        - capabilityClass
        - clusterMemberTypes
        - connectionTypes
        - id
        - isCluster
        - logo
        - name
        - type
      type: object
      allOf:
        - "$ref": "#/components/schemas/Capability"
    BasicResource:
      required:
        - clusterMember
        - id
      type: object
      properties:
        id:
          type: string
          format: uuid
        assetId:
          type: string
        location:
          "$ref": "#/components/schemas/Location"
        ip:
          type: string
        hostname:
          type: string
        firmwareVersion:
          type: string
        project:
          type: string
        clusterMember:
          type: boolean
        capabilityServices:
          type: array
          items:
            oneOf:
              - "$ref": "#/components/schemas/CapabilityService"
              - "$ref": "#/components/schemas/MultiHostCapabilityService"
              - "$ref": "#/components/schemas/SingleHostCapabilityService"
        remoteAccessService:
          "$ref": "#/components/schemas/RemoteAccessService"
    Capability:
      required:
        - actions
        - capabilityClass
        - clusterMemberTypes
        - connectionTypes
        - id
        - isCluster
        - logo
        - name
        - type
      type: object
      properties:
        id:
          type: string
          format: uuid
        capabilityClass:
          type: string
        name:
          type: string
        logo:
          type: string
        type:
          type: array
          items:
            "$ref": "#/components/schemas/CapabilityType"
        actions:
          type: object
          additionalProperties:
            oneOf:
              - "$ref": "#/components/schemas/AwxAction"
        healthCheck:
          "$ref": "#/components/schemas/CapabilityHealthCheck"
        clusterMemberTypes:
          type: array
          items:
            "$ref": "#/components/schemas/ClusterMemberType"
        connection:
          "$ref": "#/components/schemas/ConnectionType"
        executionEnvironment:
          "$ref": "#/components/schemas/ExecutionEnvironment"
        isCluster:
          type: boolean
        connectionTypes:
          uniqueItems: true
          type: array
          items:
            "$ref": "#/components/schemas/ConnectionType"
      discriminator:
        propertyName: capabilityClass
    CapabilityService:
      required:
        - ID
        - capabilityServiceClass
        - customMeta
        - managed
        - serviceClass
        - serviceMeta
        - status
        - tags
      type: object
      properties:
        service:
          type: string
        tags:
          type: array
          items:
            type: string
        serviceMeta:
          type: object
          additionalProperties:
            type: string
        capability:
          oneOf:
            - "$ref": "#/components/schemas/BaseConfigurationCapability"
            - "$ref": "#/components/schemas/DeploymentCapability"
            - "$ref": "#/components/schemas/VirtualizationCapability"
        status:
          "$ref": "#/components/schemas/CapabilityServiceStatus"
        serviceClass:
          type: string
        customMeta:
          type: object
          additionalProperties:
            type: string
        managed:
          type: boolean
        ID:
          type: string
          format: uuid
        Address:
          type: string
        Port:
          type: integer
          format: int32
        capabilityServiceClass:
          type: string
      discriminator:
        propertyName: capabilityServiceClass
    CapabilityServiceStatus:
      type: string
      enum:
        - INSTALL
        - READY
        - UNINSTALL
        - FAILED
        - UNKNOWN
    Credential:
      required:
        - credentialClass
      type: object
      properties:
        credentialClass:
          type: string
      discriminator:
        propertyName: credentialClass
    CredentialUsernamePassword:
      required:
        - password
        - username
      type: object
      allOf:
        - "$ref": "#/components/schemas/Credential"
        - type: object
          properties:
            password:
              type: string
            username:
              type: string
    DeploymentCapability:
      required:
        - actions
        - capabilityClass
        - clusterMemberTypes
        - connectionTypes
        - id
        - isCluster
        - logo
        - name
        - supportedDeploymentTypes
        - type
      type: object
      allOf:
        - "$ref": "#/components/schemas/Capability"
        - type: object
          properties:
            supportedDeploymentTypes:
              type: array
              items:
                "$ref": "#/components/schemas/DeploymentType"
    MultiHostCapabilityService:
      required:
        - ID
        - customMeta
        - managed
        - mapOfNodeIdsAndCatalogServices
        - serviceClass
        - serviceMeta
        - status
        - tags
      type: object
      allOf:
        - "$ref": "#/components/schemas/CapabilityService"
        - type: object
          properties:
            memberMapping:
              type: object
              additionalProperties:
                type: string
            mapOfNodeIdsAndCatalogServices:
              type: object
              additionalProperties:
                "$ref": "#/components/schemas/Service"
    RemoteAccessService:
      required:
        - ID
        - serviceMeta
        - tags
      type: object
      properties:
        service:
          type: string
        tags:
          type: array
          items:
            type: string
        serviceMeta:
          type: object
          additionalProperties:
            type: string
        credential:
          oneOf:
            - "$ref": "#/components/schemas/CredentialUsernamePassword"
        connectionType:
          "$ref": "#/components/schemas/ConnectionType"
        ID:
          type: string
          format: uuid
        Address:
          type: string
        Port:
          type: integer
          format: int32
    Service:
      required:
        - ID
        - Meta
        - Tags
      type: object
      properties:
        ID:
          type: string
          format: uuid
        Address:
          type: string
        Service:
          type: string
        Tags:
          type: array
          items:
            type: string
        Port:
          type: integer
          format: int32
        Meta:
          type: object
          additionalProperties:
            type: string
    SingleHostCapabilityService:
      required:
        - ID
        - customMeta
        - managed
        - serviceClass
        - serviceMeta
        - status
        - tags
      type: object
      allOf:
        - "$ref": "#/components/schemas/CapabilityService"
        - type: object
          properties:
            consulNodeId:
              type: string
              format: uuid
    VirtualizationCapability:
      required:
        - actions
        - capabilityClass
        - clusterMemberTypes
        - connectionTypes
        - id
        - isCluster
        - logo
        - name
        - type
      type: object
      allOf:
        - "$ref": "#/components/schemas/Capability"
...
  securitySchemes:
    spring_oauth:
      type: oauth2
      description: Oauth2 flow
      flows:
        authorizationCode:
          authorizationUrl: https://slm.local/auth/realms/fabos/protocol/openid-connect/auth
          tokenUrl: https://slm.local/auth/realms/fabos/protocol/openid-connect/token
          refreshUrl: https://slm.local/auth/realms/fabos/protocol/openid-connect/token
          scopes: {}
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT