# resource_management/resource_management.service/resource_management.service.rest/src/main/java/org/eclipse/slm/resource_management/service/rest/resources/CreateResourceRequest.kt
package org.eclipse.slm.resource_management.service.rest.resources

import org.eclipse.slm.resource_management.service.rest.resources.aas.submodels.digitalnameplate.DigitalNameplateV3

class CreateResourceRequest() {
    var resourceHostname: String? = null
    var resourceIp: String? = null
    var digitalNameplateV3: DigitalNameplateV3?  = null
}